Update minimum k8s version
Hopefully, this is made in a way that will make adjustable in the future
without also having to adjust the Prow configuration. And it allows for
adding additional versions to the test Matrix, in case that is useful.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 902cd4945447f843b84b84b54a7111c64498ad8f)