Emit template-error if template already exists
Previously, if a Policy defined a template that already exists, it
assumed that another Policy had defined it, and tried to get its owner
reference in order to give a helpful error message. But if the template
was created outside of a Policy, then this process would panic.

Refs:
 - https://issues.redhat.com/browse/ACM-3416

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit c07ac53f79d3d44cd17d722cc4b2697a21001eed)