Reset the complianceState when Pending
The situation this is meant to fix is not fully understood. But in case
a policy template that was previously pending somehow has some status
that has not yet made its way into the Policy's status, this will reset
it, so that the policy controller knows to send a new compliance event.

Refs:
 - https://issues.redhat.com/browse/ACM-4699

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 38b97a54f98cac17935a47f66b61fb11fd8885d1)